LIBDRUM-701. Move updateCommunityGroup to comcol-data service.
https://issues.umd.edu/browse/LIBDRUM-701
