+++ src/app/core/data/comcol-data.service.ts
@@ -29,6 +29,14 @@ import { RemoteDataBuildService } from '../cache/builders/remote-data-build.serv
 import { NotificationsService } from '../../shared/notifications/notifications.service';
 import { DSOChangeAnalyzer } from './dso-change-analyzer.service';
 import { Operation } from 'fast-json-patch';
+// UMD Customization for LIBDRUM-701
+import { CommunityGroup } from '../shared/community-group.model';
+import { CommunityGroupDataService } from './community-group-data.service';
+import { HttpOptions } from '../dspace-rest/dspace-rest.service';
+import { PutRequest } from './request.models';
+import { sendRequest } from '../shared/request.operators';
+import { HttpHeaders } from '@angular/common/http';
+// End UMD Customization for LIBDRUM-701
 
 export abstract class ComColDataService<T extends Community | Collection> extends IdentifiableDataService<T> implements CreateData<T>, FindAllData<T>, SearchData<T>, PatchData<T>, DeleteData<T> {
   private createData: CreateData<T>;
@@ -46,6 +54,8 @@ export abstract class ComColDataService<T extends Community | Collection> extend
     protected comparator: DSOChangeAnalyzer<T>,
     protected notificationsService: NotificationsService,
     protected bitstreamDataService: BitstreamDataService,
+    protected cgService?: CommunityGroupDataService, // UMD Customization for LIBDRUM-701
+
   ) {
     super(linkPath, requestService, rdbService, objectCache, halService);
 
@@ -268,4 +278,40 @@ export abstract class ComColDataService<T extends Community | Collection> extend
   public deleteByHref(href: string, copyVirtualMetadata?: string[]): Observable<RemoteData<NoContent>> {
     return this.deleteData.deleteByHref(href, copyVirtualMetadata);
   }
+
+
+
+  // UMD Customization for LIBDRUM-701
+  /**
+   * Set the community group of a community
+   * @param community
+   * @param community group
+   */
+  updateCommunityGroup(community: Community, cgId: string): Observable<RemoteData<Community>> {
+    const requestId = this.requestService.generateRequestId();
+    const communityHref$ = this.getBrowseEndpoint().pipe(
+      map((href: string) => `${href}/${community.id}`),
+      switchMap((href: string) => this.halService.getEndpoint('communityGroup', href))
+    );
+    const cgHref$ = this.cgService.getBrowseEndpoint().pipe(
+      map((href: string) => `${href}/${cgId}`)
+    );
+    observableCombineLatest([communityHref$, cgHref$]).pipe(
+      map(([communityHref, cgHref]) => {
+        const options: HttpOptions = Object.create({});
+        let headers = new HttpHeaders();
+        headers = headers.append('Content-Type', 'text/uri-list');
+        options.headers = headers;
+        return new PutRequest(requestId, communityHref, cgHref
+          , options);
+      }),
+      sendRequest(this.requestService),
+      take(1)
+    ).subscribe(() => {
+      this.requestService.removeByHrefSubstring(community.self + '/communityGroup');
+    });
+
+    return this.rdbService.buildFromRequestUUID(requestId);
+  }
+  // End UMD Customization for LIBDRUM-701
 }
